修复：拆分表格时删除最后一行导致整个表格被删除【Update TableOperate.ts】
修复：拆分表格时删除最后一行导致整个表格被删除

+++ src/editor/core/draw/particle/table/TableOperate.ts
@@ -356,8 +356,8 @@ export class TableOperate {
     const trList = element.trList!
     const curTr = trList[trIndex!]
     const curTdRowIndex = curTr.tdList[tdIndex!].rowIndex!
-    // 如果是最后一行，直接删除整个表格
-    if (trList.length <= 1) {
+    // 如果是最后一行，直接删除整个表格（如果是拆分表格按照正常逻辑走）
+    if (trList.length <= 1 && !element.pagingIndex) {
       this.deleteTable()
       return
     }
